feat: add sidebar navigation components and data table
- Implemented `DataTable` component with drag-and-drop functionality for rows.
- Created `NavMain`, `NavSecondary`, and `NavUser` components for sidebar navigation.
- Added `SectionCards` component to display key metrics.
- Introduced `SiteHeader` component for the admin dashboard layout.
- Developed `useLogout` hook for handling user logout functionality.
- Removed old `SiteHeader` component to streamline the codebase.
