+++ app/admin/courses/page.tsx
@@ -1,0 +1,3 @@
+export default function CoursesPage() {
+  return <div>Courses Page</div>;
+}
