+++ app/admin/page.tsx
@@ -1,0 +1,16 @@
+import { ChartAreaInteractive } from "@/components/sidebar/chart-area-interactive";
+import { DataTable } from "@/components/sidebar/data-table";
+import { SectionCards } from "@/components/sidebar/section-cards";
+import data from "./data.json";
+
+export default function AdminIndexPage() {
+  return (
+    <div>
+      <SectionCards />
+      <div className="px-4 lg:px-6">
+        <ChartAreaInteractive />
+      </div>
+      <DataTable data={data} />
+    </div>
+  );
+}
