+++ components/sidebar/chart-area-interactive.tsx
