+++ components/sidebar/nav-user.tsx
@@ -1,0 +1,129 @@
+"use client";
+
+import {
+  IconDashboard,
+  IconDotsVertical,
+  IconLogout,
+} from "@tabler/icons-react";
+
+import { Avatar, AvatarFallback, AvatarImage } from "@/components/ui/avatar";
+import {
+  DropdownMenu,
+  DropdownMenuContent,
+  DropdownMenuGroup,
+  DropdownMenuItem,
+  DropdownMenuLabel,
+  DropdownMenuSeparator,
+  DropdownMenuTrigger,
+} from "@/components/ui/dropdown-menu";
+import {
+  SidebarMenu,
+  SidebarMenuButton,
+  SidebarMenuItem,
+  useSidebar,
+} from "@/components/ui/sidebar";
+import { useLogout } from "@/hooks/use-logout";
+import { authClient } from "@/lib/auth-client";
+import { HomeIcon, Tv2 } from "lucide-react";
+import Link from "next/link";
+
+export function NavUser() {
+  const { isMobile } = useSidebar();
+  const { data: session, isPending } = authClient.useSession();
+  const { logout } = useLogout();
+
+  if (isPending) return null;
+
+  return (
+    <SidebarMenu>
+      <SidebarMenuItem>
+        <DropdownMenu>
+          <DropdownMenuTrigger asChild className="cursor-pointer">
+            <SidebarMenuButton
+              size="lg"
+              className="data-[state=open]:bg-sidebar-accent data-[state=open]:text-sidebar-accent-foreground"
+            >
+              <Avatar className="h-8 w-8 rounded-lg">
+                <AvatarImage
+                  src={
+                    session?.user.image ||
+                    `https://avatar.vercel.sh/${session?.user.email}`
+                  }
+                  alt={session?.user.name}
+                />
+                <AvatarFallback className="rounded-lg">CN</AvatarFallback>
+              </Avatar>
+              <div className="grid flex-1 text-left text-sm leading-tight">
+                <span className="truncate font-medium">
+                  {session?.user.name || session?.user.email.split("@")[0]}
+                </span>
+                <span className="text-muted-foreground truncate text-xs">
+                  {session?.user.email}
+                </span>
+              </div>
+              <IconDotsVertical className="ml-auto size-4" />
+            </SidebarMenuButton>
+          </DropdownMenuTrigger>
+          <DropdownMenuContent
+            className="w-(--radix-dropdown-menu-trigger-width) min-w-56 rounded-lg"
+            side={isMobile ? "bottom" : "right"}
+            align="end"
+            sideOffset={4}
+          >
+            <DropdownMenuLabel className="p-0 font-normal">
+              <div className="flex items-center gap-2 px-1 py-1.5 text-left text-sm">
+                <Avatar className="h-8 w-8 rounded-lg">
+                  <AvatarImage
+                    src={
+                      session?.user.image ||
+                      `https://avatar.vercel.sh/${session?.user.email}`
+                    }
+                    alt={session?.user.name}
+                  />
+                  <AvatarFallback className="rounded-lg">
+                    {session?.user.name?.charAt(0).toUpperCase() ||
+                      session?.user.email?.charAt(0).toUpperCase()}
+                  </AvatarFallback>
+                </Avatar>
+                <div className="grid flex-1 text-left text-sm leading-tight">
+                  <span className="truncate font-medium">
+                    {session?.user.name || session?.user.email.split("@")[0]}
+                  </span>
+                  <span className="text-muted-foreground truncate text-xs">
+                    {session?.user.email}
+                  </span>
+                </div>
+              </div>
+            </DropdownMenuLabel>
+            <DropdownMenuSeparator />
+            <DropdownMenuGroup>
+              <DropdownMenuItem asChild className="cursor-pointer">
+                <Link href="/" className="flex items-center gap-2">
+                  <HomeIcon />
+                  Homepage
+                </Link>
+              </DropdownMenuItem>
+              <DropdownMenuItem asChild className="cursor-pointer">
+                <Link href="/admin" className="flex items-center gap-2">
+                  <IconDashboard />
+                  Dashboard
+                </Link>
+              </DropdownMenuItem>
+              <DropdownMenuItem asChild className="cursor-pointer">
+                <Link href="/admin/courses" className="flex items-center gap-2">
+                  <Tv2 />
+                  Courses
+                </Link>
+              </DropdownMenuItem>
+            </DropdownMenuGroup>
+            <DropdownMenuSeparator />
+            <DropdownMenuItem onClick={logout} className="cursor-pointer">
+              <IconLogout />
+              Log out
+            </DropdownMenuItem>
+          </DropdownMenuContent>
+        </DropdownMenu>
+      </SidebarMenuItem>
+    </SidebarMenu>
+  );
+}
