+++ hooks/use-logout.ts
@@ -1,0 +1,25 @@
+"use client";
+
+import { authClient } from "@/lib/auth-client";
+import { useRouter } from "next/navigation";
+import { toast } from "sonner";
+
+export function useLogout() {
+  const router = useRouter();
+
+  async function logout() {
+    await authClient.signOut({
+      fetchOptions: {
+        onSuccess: () => {
+          toast.success("Logged out successfully");
+          router.push("/");
+        },
+        onError: () => {
+          toast.error("Failed to log out");
+        },
+      },
+    });
+  }
+
+  return { logout };
+}
